Properly handle breakpoints and implement p, t, and g correctly
When the target is stopped, we don't want our breakpoints to pollute
the memory space. Therefore we should be inserting breakpoints when
we resume execution and removing them when we stop again. Also, if we
stop due to a breakpoint, we need to backtrack 1 byte to execute the
missed instruction.